requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.0
redis==5.0.0
celery==5.4.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.0
httpx==0.27.0
oss2==2.18.0
alibabacloud_dysmsapi20170525==2.0.24
pytest==8.3.0
pytest-asyncio==0.23.0
flake8==7.1.0
mypy==1.11.0
